+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/block/RainShieldBlock.kt
@@ -8,8 +8,8 @@ import net.minecraft.world.level.BlockGetter
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.BaseEntityBlock
 import net.minecraft.world.level.block.Block
+import net.minecraft.world.level.block.Blocks
 import net.minecraft.world.level.block.EntityBlock
-import net.minecraft.world.level.block.SoundType
 import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.entity.BlockEntityTicker
 import net.minecraft.world.level.block.entity.BlockEntityType
@@ -17,12 +17,12 @@ import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.level.block.state.StateDefinition
 import net.minecraft.world.level.block.state.properties.BlockStateProperties
 import net.minecraft.world.level.block.state.properties.BooleanProperty
+import net.minecraft.world.phys.shapes.CollisionContext
+import net.minecraft.world.phys.shapes.VoxelShape
 
 class RainShieldBlock : EntityBlock, Block(
-	Properties
-		.of()
-		.sound(SoundType.STONE)
-		.strength(2f)
+	Properties.ofFullCopy(Blocks.END_ROD)
+		.lightLevel { 0 }
 ) {
 
 	init {
@@ -38,6 +38,10 @@ class RainShieldBlock : EntityBlock, Block(
 
 	override fun canConnectRedstone(state: BlockState, level: BlockGetter, pos: BlockPos, direction: Direction?): Boolean {
 		return true
+	}
+
+	override fun getShape(state: BlockState, level: BlockGetter, pos: BlockPos, context: CollisionContext): VoxelShape {
+		return SHAPE
 	}
 
 	override fun neighborChanged(
@@ -70,6 +74,7 @@ class RainShieldBlock : EntityBlock, Block(
 
 	companion object {
 		val ENABLED: BooleanProperty = BlockStateProperties.ENABLED
+		val SHAPE: VoxelShape = box(6.0, 0.0, 6.0, 10.0, 16.0, 10.0)
 	}
 
 }